ChatbotAgent/backend/requirements.txt
fastapi
uvicorn
gunicorn
langchain-core
langchain-community
langchain-openai
python-dotenv

